🚚 Rename:  Collapse, Slide, Scale, Fade 폴더 전부 Transition 폴더 밖으로 배치 & Transitions -> Transition으로 폴더명 변경
