+++ src/components/general/Collapse/Collapse.mdx
@@ -11,7 +11,7 @@ Expand(확장)/Collapse(축소) Transition<br />
 ## Import
 
 ```js
-import { Collapse } from '@/components/general/Transitions';
+import { Collapse } from '@/components/general/Collapse';
 ```
 
 ## Usage
