+++ src/components/general/Fade/Fade.mdx
@@ -11,7 +11,7 @@ Fade-In/Fade-Out Transition<br />
 ## Import
 
 ```js
-import { Fade } from '@/components/general/Transitions';
+import { Fade } from '@/components/general/Fade';
 ```
 
 ## Usage
