+++ src/components/general/Fade/index.ts
