+++ src/components/general/Scale/Scale.mdx
@@ -11,7 +11,7 @@ Scale-In/Scale-Out Transition<br />
 ## Import
 
 ```js
-import { Scale } from '@/components/general/Transitions';
+import { Scale } from '@/components/general/Scale';
 ```
 
 ## Usage
