+++ src/components/general/Slide/Slide.hooks.ts
@@ -1,7 +1,7 @@
 import { useRef, useState, useCallback } from 'react';
 import useJinni from '@/hooks/useJinni';
 import { isNumber } from '@/utils/isNumber';
-import { flatDuration } from '../Transition.utils';
+import { flatDuration } from '@/components/general/Transition';
 import { SlideProps } from './Slide';
 
 type useSlideProps = {
