+++ src/components/general/Slide/Slide.mdx
@@ -11,7 +11,7 @@ Slide-In/Slide-Out Transition<br />
 ## Import
 
 ```js
-import { Slide } from '@/components/general/Transitions';
+import { Slide } from '@/components/general/Slide';
 ```
 
 ## Usage
