+++ src/components/general/Slide/index.ts
