+++ src/components/general/Transition/Transition.mdx
@@ -3,15 +3,15 @@ import * as TransitionStories from './Transition.stories';
 
 <Meta of={TransitionStories} />
 
-# Transitions
+# Transition
 
-Transitions은 UI를 좀 더 풍부하고 사용하기 쉽게 만들어 줌<br />
+Transition은 UI를 좀 더 풍부하고 사용하기 쉽게 만들어 줌<br />
 (Collapse, Fade, Slide, Scale 컴포넌트는 Transition 컴포넌트를 기반으로 만들어짐)
 
 ## Import
 
 ```js
-import { Transition } from '@/components/general/Transitions';
+import { Transition } from '@/components/general/Transition';
 ```
 
 ## Usage
