+++ src/components/general/Transition/Transition.scss
