+++ src/components/general/Transition/index.ts
@@ -1,0 +1,4 @@
+export { default as Transition, DEFAULT_DURATION } from './Transition';
+export type { TransitionProps } from './Transition';
+export { flatDuration } from './Transition.utils';
+export { TRANSITION_ARG_TYPES } from './Transition.constants';
