--- src/components/general/Transitions/index.ts
@@ -1,6 +1,0 @@
-export { default as Transition } from './Transition';
-export type { TransitionProps } from './Transition';
-export { Fade } from './Fade';
-export { Collapse } from './Collapse';
-export { Slide } from './Slide';
-export { Scale } from './Scale';
